updatrw

+++ view/visualisador.py
@@ -48,7 +48,10 @@ st.write("l")
 df_final["prova falta"] = df_final["nota"].isna()
 # calcula os dados necessarios
 groupby_trimestre_falta = df_final.groupby('trimestre').agg({'nota':"sum",'prova falta': 'sum'}).reset_index()
-groupby_trimestre_falta["quanto falta trimestre"] = int(18-groupby_trimestre_falta["nota"])/groupby_trimestre_falta["prova falta"]
+#groupby_trimestre_falta["quanto falta trimestre"] = int(18-groupby_trimestre_falta["nota"])/groupby_trimestre_falta["prova falta"]
+groupby_trimestre_falta["quanto falta trimestre"] = (
+    ((18 - groupby_trimestre_falta["nota"]) / groupby_trimestre_falta["prova falta"]).astype(int)
+)
 groupby_trimestre_falta["media trimestre"] = groupby_trimestre_falta["nota"]/3 
 groupby_trimestre_falta["media trimestre"] =  groupby_trimestre_falta["media trimestre"].round(1)
